Allows for deletion of walls when cursor pulled back
ERROR: only one wall is deleted. Instead of re-enabling the previous wall, a new wall is placed then turns to face the cursor (overlapping the previous wall). When next deletion attempt is made, since the upper and lower extenders are reversed, the wall tries to build into the next wall and room is declared done